forgejo: Split API operations by resource group
Move the Forgejo gateway API surface out of one declaration-kind
file and into files grouped by operation family. Pull requests,
reviews, issue comments, statuses, content/templates, labels,
users, repositories, assignees, and shared pagination/path helpers
now keep endpoint methods close to their supporting option and DTO
types.

This is a package-local organization change. It preserves exported
method names, type names, JSON tags, endpoint paths, query
parameters, and transport behavior.
